Mostly GUIs update
Added a file selector class that allows you to pick a file.
Added Hanzi.java which is a class for individual Chinese characters. It's implementation is more-or-less the same as Word.java, but there are minor differences, and my justification is that the abstraction is different. Can probably help with checking for good/bad input later.
Added Study.java which is a GUI for studying a given set of vocab words or individual chinese characters.
Also added a directory for storing and segregating the data.
GUI.java was updated to now also have a field which displays the vocab words or hanzi already in a set (as of now, it's on the user to make sure that sets have unique elements).
I think this was after the last update, but conversions can now handle "strings" of pinyin/zhuyin syllables and convert whole transcription sentences.
Word.java now has a method to get a formatted version of the English translation data (although as of now, it's only meant to be used for GUIs).