Major Study.java update
mainly updated the functionality of Study.java. Now optionally displays the transcriptions, and when initialized to not show them, pressing the buttons will. When the next word is reached, the transcriptions will disappear when the showTranscription boolean is initialized to false.
Counters now update appropriately, and answer checking is case insensitive.
Support has now been added to study either Word objects or Hanzi objects.